Moved GitHub client methods out to their proper package and broke the case classes out accordingly.
