+++ app/com/ardusjax/webservice/clients/ArtifactoryClient/ArtifactoryWSC.scala
@@ -1,0 +1,17 @@
+package com.ardusjax.webservice.clients.ArtifactoryClient
+
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import play.api.libs.json._
+import play.api.libs.ws.WSClient
+
+import scala.concurrent.Future
+
+/**
+  * Created by Christopher G. O'Brian on 12/5/15.
+  */
+class ArtifactoryWSC(ws: WSClient, baseUrl: String) {
+
+
+
+
+}
