+++ app/controllers/Application.scala
@@ -1,8 +1,10 @@
 package controllers
 
 import javax.inject.Inject
+import com.ardusjax.webservice.clients.ArtifactoryClient.ArtifactoryWSC
+import com.ardusjax.webservice.clients.GitHubClient.GitHubWSC
+
 import scala.concurrent.duration._
-import com.ardusjax.webservice.clients.ArtifactoryWSC
 import org.joda.time.Seconds
 import play.api._
 import play.api.libs.ws.{WS, WSClient}
@@ -14,10 +16,9 @@ import scala.concurrent.Await
 object Application  extends Controller {
   def index = Action {
     def githubclient = {
-      new ArtifactoryWSC(WS.client, "https://api.github.com")
+      new GitHubWSC(WS.client, "https://api.github.com")
     }
     val result = Await.result(githubclient.repositories(), 10.seconds )
     Ok(views.html.index(result))
   }
-
 }
